[REFACTOR] components/dashboard 하위 컴포넌트 리팩토링
