+++ components/Dashboard/@index.ts
@@ -1,5 +1,5 @@
-import ComprehensiveData from './ComprehensiveData';
-import OrderData from './OrderData';
-import Graph from './Graph';
+import VisitSection from 'components/Dashboard/VisitSection';
+import ProductStatisticsSection from 'components/Dashboard/ProductStatisticsSection';
+import ComprehensiveDataSection from './ComprehensiveDataSection';
 
-export { ComprehensiveData, OrderData, Graph };
+export { ComprehensiveDataSection, VisitSection, ProductStatisticsSection };
