--- components/Dashboard/GraphBody.tsx
@@ -1,15 +1,0 @@
-import styled from '@emotion/styled';
-
-const GraphBody = () => {
-  return <S.Container></S.Container>;
-};
-
-const S = {
-  Container: styled.div`
-    width: 100%;
-    height: 448px;
-    border: 1px solid black;
-  `,
-};
-
-export default GraphBody;
