+++ components/Dashboard/common/Shadow.tsx
@@ -1,31 +1,18 @@
 import styled from '@emotion/styled';
-import { BOX_SIZES } from 'constants/dashboard';
-import React, { ReactNode } from 'react';
 
-interface BoxSize {
-  boxSize: 'SMALL' | 'MEDIUM' | 'LARGE';
-}
+import type { WrapProps } from 'typings/wrapperProps';
 
-interface ShadowProps extends BoxSize {
-  children: ReactNode;
-}
+const Shadow = ({ children }: WrapProps) => {
+  const S = {
+    Background: styled.div`
+      background: #ffffff;
 
-const Shadow = ({ children, boxSize }: ShadowProps) => {
-  return <S.Background boxSize={boxSize}>{children}</S.Background>;
-};
+      box-shadow: 4px 4px 23px rgba(0, 0, 0, 0.14);
+      border-radius: 15px;
+    `,
+  };
 
-const S = {
-  Background: styled.div<BoxSize>`
-    position: relative;
-
-    width: ${({ boxSize }) => BOX_SIZES[boxSize].width}px;
-    height: ${({ boxSize }) => BOX_SIZES[boxSize].height}px;
-
-    background: #ffffff;
-
-    box-shadow: 4px 4px 23px rgba(0, 0, 0, 0.14);
-    border-radius: 15px;
-  `,
+  return <S.Background>{children}</S.Background>;
 };
 
 export default Shadow;
